Reverse order of board and layout pre-image hook
It is envisaged that a board pre-image hook will never need to generate
the genimage input config. However a board pre-image hook may need to
install files into the rootfs before image creation.

While it is likely that a post-build hook will be introduced, reversing
the order of the pre-image hook at least allows board specific fixups
to be performed (eg on the rootfs) before the image is generated.