meta: Use signed-by for RPi apt URIs
In accordance with Debian best practices for third-party repos [1] use
signed-by to specify the key used to authenticate the Release files for
RPi apt sources.

This key is currently installed by package raspberrypi-archive-keyring.

Relocate the layer describing the RPi apt sources as it's not arm64
specific (same source is used in Raspian).

[1] https://wiki.debian.org/DebianRepository/UseThirdParty