Enhance dependency checking
Allow multiple files containing dependencies to be specified. The
core base dependency list will always be included. This may be useful
for examples that require additional packages compared to the core
dependencies.

Use dpkg as an additional test to decide if a pkg needs to be
installed. Packages that provide only libraries fail the hash test
regardless of whether they are installed or not, so they need to be
checked via other means. For example libavcodec-dev.

A failure to install or check deps must exit with error.