Support nested images and fine tune output dirs.
In order to support 'nested' image creation with genimage (eg, where an
image can be included in another), the input and output directory args
passed to genimage need to be the same.
Relocate the deploy dir so it's outside the artefact dir.
Update the documentation to match, including some small additional
changes.