Add apt keydir support
Being able to provide APT with a set of keys to use for pkg download
at chroot creation has a few advantages:

It keeps rpi-image-gen relatively agnostic to what's installed
on the host, ie we don't need the host keyring to contain keys that
we know are mandatory for bootstrapping (eg for archive.raspberrypi.com).

It means the build should work on any Debian-ish system, not just RPi
OS 64-bit (where raspberrypi-archive-keyring is installed by default).

If the user requires a particular set of keys for their configuration,
they can provide them.

The trade-off is that we can't assume we can bootstrap from a
signed-by=/usr/share/keyrings/xxx mirror (ie because we can't assume
the keys for archive.raspberrypi.com will exist in /usr/share/keyrings/
on the host). Instead, after installing all packages, we test that the
required signed-by key exists in the chroot then add an entry in
<chroot>/etc/apt/sources.list.d/ so that the mirror will be signed-by
qualified when the image is executing on the target.

The keydir is propagated into the chroot by mmdebstrap setting
Dir::Etc::TrustedParts. This needs to be removed since it's only valid
at chroot creation. Removing it in the essential hook means it's
possible to execute apt operations within the chroot, eg at customize,
if required.

Because Dir::Etc::TrustedParts defines the keys APT will use, retaining
this in the image would render all mirrors unable to be verified (ie
because the keydir is not accessible from within the chroot).

Remove the manual installation of the raspberrypi key into the chroot.
This is redundant because it's installed there by the
raspberrypi-archive-keyring pkg.