Add native 32-bit raspbian support
Add the raspbian key to the keydir so it's included in the default APT
key list used to create the chroot.

Add new meta layers for:
 raspbian bookworm base
 armhf specific v7 kernel+initramfs

Add new profile which can be used to create a native 32bit system with
v7 kernel, rpi utils and systemd wired networking.

Add new config to use this profile to create an image with simple
partition layout similar to that of RPi OS.