meta: 64-bit kernels must specify arch
Failure to specify the arch in this way will lead to these packages
failing to download if used with a native 32-bit userland base (eg
from raspbian). This isn't so much of a problem for the 2712 kernel
because at the current time, binaries from raspbian will not execute
on it. It is, however, a problem if there is a need to use the v8
kernel with raspbian userland.