Simplify APT proxy usage
An HTTP APT proxy can be passed directly to bdebstrap via
--aptopt Acquire::http so there is no need to implement it via YAML.
The YAML proxy layer was a bad design choice anyway because it had
to be included the profile in order for the proxy to be used, which
was unintuitive.
Remove the obsolete meta layer and pass the proxy via aptopt instead.

Add a new cleanup hook to remove the proxy from the chroot.