Add SBOM generation
Hook up SBOM generation using syft to the end of the post-build stage so
it takes place immediately before genimage invocation. The output format
of the SBOM is configurable. If syft is not installed on the host it
will be downloaded to a location which will be prefixed to the current
PATH.
Update documentation.