post-build env additions
s/ENV_IMAGE/ENV_POST_BUILD/g
This is a more appropriate name because this variable actually holds
the environment for everything after rootfs build which can encompass
post-build and pre-image hooks, sbom, image creation, etc.

Include additional variables in the post-build env which provide
full paths to locations for the chosen board, profile and image
directories.

Pass the post-build env to genimage so image-gen variables are
available in its execution env. This may be beneficial to genimage
config exec-pre and exec-post hooks.