Exit after post-build if only generating rootfs
It's important that the post-build hooks run if only performing rootfs
generation because they may be needed to hook into actions that perform
operations on the rootfs (e.g. calling out to an external build system
to install into the rootfs). If rpi-image-gen is being used to generate
a rootfs with no output image (e.g. to prepare an SDK), invoking
post-build hooks before exiting may be critical functionality that's
required.