Extend bdebstrap hooks to image and board
An image or board may need to utilise bdebstrap hooks rather than being
limited to a post-build hook. The bdebstrap hooks execute in the context
of bdebstrap invocation, which mean they operate on the chroot as it is
being created and at the respective stage (i.e. setup, essential,
customize and cleanup). Adding support of these hooks to image and board
broadens the scope of what can be performed outside of the meta layers
and enables further customisation options.

One example of why these hooks may be useful would be if an image layout
wanted to install files in the initramfs that were required for specific
reasons. It would be simpler to install the files in the rootfs using a
customize hook during chroot creation rather than afterwards at
post-build.