Add --force to bdebstrap args
If --force is used, bdebstrap will not stop if the output directory
already exists, which is desirable behaviour for rebuilding. However,
unless told otherwise, mmdebstrap will error out if the output
directory exists. This means it's not possible to rebuild/restart a
a build on top of previous output. While an inconvenience, it's the
understanding that this is the default behaviour of mmdebstrap, which
inherits some behaviour from debootstrap [1].

mmdebstrap supports -skip=check/empty which side-steps the check of
an existing directory and allows a rebuild to be done. Furthermore, by
allowing a rebuild there is cache state replay which prevents all
packages from being download again, therefore speeding up (re)build time
significantly.

Unfortunately, passing -skip=check/empty from YAML isn't available in
bdebstrap until v0.6.0, and this version is much more recent that the
version shipping in bookworm. But by allowing bdebstrap to proceed we
open the first gate.

Refs:
[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=833525