Add framework for device and image options
Device and image layout directories can provide a file to define
the configuration variables they support. The names of these files
are themselves configurable at the top level.

If a file exists, it is loaded at the parameter assembly stage and
serves to set the default value of a variable. Default values are
only set if the variable has not been set prior (eg via the config).

It is envisaged that these option files could serve as an input to
auto-generation of the documentation for the device or image layout,
along with a template. This may help with maintenance as more options
are added.

Create a new options file for both in-tree image layouts to hold
default values (currently only partition sizes). Fixup the AB layout to
reference the partition sizes from the options rather than statically
within the pre-image hook.

Add new load function in scripts to read and aggregate the settings
from these files before loading user options (user options can always
override anything). The aggregation includes a prefix which ensures the
variable is limited to the intended configuration scope.

Misc:
 Bug fix setting default image name.
 Bug fix reading user options