Add env helper
This helper can be used to inspect rpi-image-gen environment
variables which may be useful in scripts, particularly those relating to
capabilties. By default it will be in PATH for all hooks to use. If
multiple variables are passed, success will only result if the
appropriate test is true for all variables inspected (ie logical AND).
Variables may be passed with or without an IGconf_ prefix. This prefix
will always be present when the variable is tested.