Add bdebstrap hook support to external directory
Similar to image and board dirs, if the optional external directory is
provided and contains a bdebstrap sub-directory, executable scripts
within it which conform to the defined naming convention will be run at
the appropriate stage(s) of chroot creation.

This may be of benefit to users who wish to utilise board and image from
in-tree, but who wish to make their own customisations to the chroot.