Another compat change
Hopefully the last.. Pedantic but worthwhile.

In the interests of keeping configuration variables as short as
possible, rename IGconf_device_user1name to IGconf_device_user1.
This user account can be granted functionality via different
variables which themselves need to be associated to this user, so
it makes sense to try and keep the identifier for the user name
as short as possible.

Fix typo in docker meta variable, and rename (usr1name -> user1).