Always fixup user1 .ssh permissions if it exists
Rather than relying on a variable only present if
net-misc/openssh-server was included in the profile to set user1's
.ssh/ permissions, just test if the directory exists instead.

This fixes a bug where user1/.ssh/authorized_keys would be left owned by
root if IGconf_meta_ssh_pubkey_user1 was not set to a non-empty string
(default). On-device, this would have rendered user1 unable to write to
their own authorized_keys file, or sshd being unable to load this file
for use (moot point as it would be empty). This didn't happen if
IGconf_meta_ssh_pubkey_user1 was set to a non-empty string.