AB: add dynamic boot slot logic and support
The chosen bootloader device-tree attributes now drive the AB boot mech
to support automatic 'slot' based rootfs mounting on systems that boot
from SD/eMMC or NVMe.

Add a template based scheme which can be used to auto-generate a slot
helper easily usable from image layouts that wish to support AB. The
helper can be used in udev rules which are installed in the initramfs
to create the device node alias symlinks for the mapped slot partition
block devices under /dev/disk/by-slot/

If an image layout is to use the slot helper it must define the slot
indices needed by the AB mech in a simple text file containing key=value
pairs to identify the partition numbers for the supported slot
components. This file dictates the slot map that will be pinned via
udev rules in the initramfs and it must match the order of the respective
partitions in the genimage template for the given image layout.

Using information directly from the bootloader in conjunction with
the defined partition map provides a 'one source of truth' method for
discovery of the active and other slots at run-time.

The installed slot helper supports several different command line
arguments and can be used for many slot related operations, including
generation of a tryboot fragment suitable for autoboot.txt using the
current boot slot partition as a reference. One example use of this
could be write out autoboot.txt after a successful tryboot in order to
commit a new update.

Misc related:
o Rename rootfs partitions to 'system'.
o Mark AB system non-critical mountpoints with nofail.
o Partition images for A or B must not be different in any way. This
  is critically important because they must be able to be written to any
  slot. While it's possible to specify the rootfs system 'slot' via
  UUID or other unique identifier on the kernel cmdline, this does not
  guarantee slot pinning and introduces more dependencies in the slot
  mgmt. Delegating the system slot at run-time using information from
  the bootloader is much preferred.

Additional misc:
o Precompute filesystem UUIDs for image generation
o Use a defined mke2fs.conf when creating ext4 filesystems