Improve initramfs hook prerequisites
This hook should only run if initramfs-tools is installed in the chroot
because we don't (currently) support initramfs generation via any other
means, eg dracut. In addition to making sure we don't invoke
update-initramfs, we also shouldn't copy in the static overlays that are
specific to initramfs-tools.

Exit this script immediately if initramfs-tools isn't installed by using
dctrl to check install presence.