python export: genimage cfg parser improvements
Rewrite the low-level libconfuse parsing to translate the genimage
config file to a dictionary in entirety. Doing this radically
simplifies higher level operations by avoiding the need to use
complicated regex to extract sections/fields. Another advantage of
this is that sub-sections are now preserved in the translation,
allowing them to be inspected or referenced if required.

Enhance the fstab based matching scheme to fallback to using a
genimage defined mountpoint to associate with an fstab entry via
fs_file [fstab(5)]. This won't work in all cases but it's a
reasonable-ish assumption make for a fallback implementation.
Identification via a unique identifier such as a filesystem UUID is
recommended because this ensures a unique key can be used at the lookup
stage and the correspond fstab line can be a guaranteed match.

Various other improvements made to the parsing code and general
readability, including addition of a new attributes section in the JSON
of which will hold global image attributes, eg size.
Version bumped accordingly.