Add image profile support
An image layout may require it's own YAML layer(s) in order to support
particular functionality. Add support for an optional profile name to
be defined and if so, load that profile from the core engine.

Introduce a scope when loading a profile which can help define how
layers within the profile are searched for. By design, the main profile
should only be able to resolve layers provided by in-tree meta, external
dir meta, or external namespace meta. In contrast, an image profile may
be able to resolve a layer from any of these locations or it's own
layout-specific meta.