fs related fixups for image layouts
Set fs_passno to 0 for BTFS filesystems, inline with docs [1].

Install packages to support the device filesystems and ensure
initramfs-tools is informed about the rootfs type so it can copy
appropriate fsck binaries into the image.

initramfs-tools uses different ways to establish the rootfs type when
FSTYPE=auto (the default). However, none of these sources can be
relied upon at the time the initramfs is built in the hook. This results
in fsck missing from the initramfs image. To combat this, if an image
layout indicates a particular rootfs type, hint it to initramfs-tools
via a core hook. If a layout does not, it must fixup the initramfs
config itself.

[1] https://btrfs.readthedocs.io/en/latest/fsck.btrfs.html