Add support for configurable target type
In some cases it may be desirable to create a tarball of the device
filesystem rather than a directory (eg for Docker image development). To
support this, add a new sys build variable which can be used to specify
a path to the target object that bdebstrap will create to hold the
device filesystem. The value of this variable is passed directly to
bdebstrap via --target and can be a path to a directory (default) or a
file. If a file, rpi-image-gen will terminate after bdebstrap execution
completes as later stages (eg hooks and genimage) only support operating
on a directory.

To support this, rather than pinning the bdebstrap hooks to operate on a
directory, pass the chroot directory from bdebstrap to the hooks instead
allowing them to operate on the chroot regardless of whether the target
is a file or a directory (if a file, a temporary directory will be used
and deleted upon bdebtrap termination).