Update webkiosk for rpt cage
The recently packaged rpt1 cage now marks xwayland as a dependency
so it doesn't need to be explicitly installed.

Note that vc4-hdmi-X will present a keyboard pointer compatible device
to the input framework if the display supports CEC (libinput
list-devices). As far as Cage is concerned, if any input device has
pointer capability, it shows the cursor [1].
FWIW Sway (another compositor) automatically hides the cursor after a
period of inactivity.

[1] https://github.com/cage-kiosk/cage/issues/299