Store actual image sizes in the JSON description
It's imperative that the actual sizes of the appropriate image files
are encapsulated in the JSON because provisioning relies on them being
accurate. The genimage config can't be relied upon to provide these
numbers, particularly as they may contain sizes defined in percent.

Augment the search of image files with an env variable to enable the
location of build artefacts to be found.

Because we now need all image files to exist in order to obtain their
size, it explicitly moves the stage of where image2json can be invoked
to post-image. It will deliberately raise errors if image files
referenced by the genimage config cannot be found.

Extend the common post-image hook to generate the layout description,
taking into account any fstab files that may be present.

Add sparse partition image generation to both in-tree image layouts so
they can be used to provisioning via fastboot and remove inclusion of
output path in the templates - it doesn't need to be there.