Add initial provisioning description support
Define a new JSON object for provisioning and use it to generate
the slot map for AB systems. This has several advantages including:

1) It can easily be parsed to generate a shell fragment to define the
    partition pinning needed by the udev slot helper.
2) It can be used directly in the image description JSON blob.

The provisionmap object is crucial for provisioning and serves as the
defacto description of how a device will be provisioned in terms of
utilisation of the partitions available in the partitionimages object.

Add initial provisioning maps for in-tree image layouts to support
encrypted and non-encrypted systems.

Using jq to read the JSON means it must be added to tool's deps.