Fix RPi apt key installation
On non Raspberry Pi OS build hosts, unless the RPi key is provided to
apt, it will fail to authenticate the repo. rpi-image-gen does cater
for this. However, the key is not accepted by apt because it's
currently stored as ascii-armored but does not have a .asc suffix.

Solutions are to rename it to have a .asc suffix, or de-armor and
rename it so it has a .gpg suffix.
The latter is preferred for wider interoperability reasons as per
apt-key(8). Doing so results in the same key that would be obtained
from (example):

gpg --keyserver keyserver.ubuntu.com --recv-keys CF8A1AF502A2AA2D763BAE7E82B129927FA3303E

Resolves #52