ssh: Fix public key only configuration
Use the correct igconf option to inspect the variable governing whether
public key only auth should be used.

Ensure igconf reports an error and exits non-zero if an unsupported
option is provided. Unfortunately, this won't trigger a build failure
because scripts in the YAML are executed in a new process, and
enforcing strict shell behaviour for the whole build is not viable.